+++ src/main/java/com/loquinecrud/loquinecrud/service/BookRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
-    List<Book> findByAuthorId(Long id);
+public interface BookRepository extends JpaRepository<Book, Integer> {
+    List<Book> findByAuthorId(Integer id);
 }
